服务层 接口层实现

+++ aspnet-core/src/Tiger.Domain/Volo/Abp/Sass/Tenants/ITenantRepository.cs
@@ -49,7 +49,7 @@ namespace Tiger.Volo.Abp.Sass.Tenants
         /// <param name="filter"></param>
         /// <param name="cancellationToken"></param>
         /// <returns></returns>
-        Task<long> GetCountAsync(
+        Task<int> GetCountAsync(
             string filter = null,
             CancellationToken cancellationToken = default);
